+++ estadificacion/__init__.py
@@ -8,6 +8,11 @@ profesional.
 """
 
 from estadificacion.builder import proponer_estadificacion
+from estadificacion.incompleta import (
+    AnalisisEstadificacionIncompleta,
+    ComponenteIndeterminado,
+    analizar_estadificacion_incompleta,
+)
 from estadificacion.models import (
     DISCLAIMER,
     ComponenteEstadio,
@@ -18,12 +23,17 @@ from estadificacion.staging_systems import (
     ComponenteDef,
     SistemaEstadificacion,
     agrupar_estadio,
+    estadios_candidatos,
     familia_de_valor,
+    orden_estadio,
     sistema_para_cancer,
 )
 
 __all__ = [
     "proponer_estadificacion",
+    "analizar_estadificacion_incompleta",
+    "AnalisisEstadificacionIncompleta",
+    "ComponenteIndeterminado",
     "PropuestaEstadificacion",
     "ComponenteEstadio",
     "DISCLAIMER",
@@ -32,5 +42,7 @@ __all__ = [
     "SistemaEstadificacion",
     "sistema_para_cancer",
     "agrupar_estadio",
+    "estadios_candidatos",
+    "orden_estadio",
     "familia_de_valor",
 ]